+++ main/src/main/scala/org/clulab/sequences/LexiconNER.scala
@@ -1,19 +1,52 @@
 package org.clulab.sequences
 
-import java.util.function.Consumer
-
 import org.clulab.processors.Sentence
-import org.clulab.struct.CompactLexiconNER
-import org.clulab.struct.DebugBooleanHashTrie
-import org.clulab.struct.IntHashTrie
-import org.clulab.struct.{BooleanHashTrie, EntityValidator, TrueEntityValidator}
-import org.clulab.utils.Files.loadStreamFromClasspath
-import org.clulab.utils.Serializer
-import org.slf4j.{Logger, LoggerFactory}
+import org.clulab.struct.{EntityValidator, TrueEntityValidator}
 
 import scala.collection.mutable
 
-abstract class LexiconNER(knownCaseInsensitives: Set[String], useLemmas: Boolean) extends Tagger[String] with Serializable {
+/**
+  * The abstract base class for several concrete child classes used for Named Entity
+  * Recognition (NER) based on the contents of lexica, which are lists of words and
+  * phrases representing named entities
+  *
+  * For all of these classes, NER labels are derived from the file names of the lexica
+  * or the records in overrideKBs by the [[org.clulab.sequences.LexiconNERBuilder LexiconNERBuilders]].
+  * This class, via variables [[org.clulab.sequences.LexiconNER.USE_FAST USE_FAST]] and
+  * [[org.clulab.sequences.LexiconNER.USE_COMPACT USE_COMPACT]], controls which builder
+  * use used.
+  *
+  * The collection of child classes is small:
+  * <ul>
+  *   <li>
+  *     The [[org.clulab.sequences.SeparatedLexiconNER SeparatedLexiconNER]] is closest
+  *     to the original implementation.  It has a [[org.clulab.struct.BooleanHashTrie BooleanHashTrie]]
+  *     for each label and in that trie, Boolean values indicate that the sequence of
+  *     strings leading there is a named entity.  Each trie structure must be searched for
+  *     potential named entities.
+  *   </li>
+  *   <li>
+  *     The [[org.clulab.sequences.CombinedLexiconNER CombinedLexiconNER]] stores instead
+  *     of the Boolean in the BooleanHashTrie an Int in an [[org.clulab.struct.IntHashTrie IntHashTrie]].
+  *     The Int indicates which of the labels is the one to use for the entity just found.
+  *     In this way, only one trie (or two if there are different case sensitivity settings)
+  *     needs to be searched no matter how many labels there are (at least until Integer.MAX_VALUE).
+  *   </li>
+  *   <li>
+  *     The [[org.clulab.sequences.CompactLexiconNER CompactLexiconNER]] uses the same
+  *     strategy to minimize the number of tries, but also converts the tries into
+  *     [[org.clulab.sequences.CompactTrie CompactTries]] which consist of arrays
+  *     of integers indicating offsets into other arrays.  In this way the time it takes
+  *     to de/serialize the NER is reduced, and some lookup operations are made more efficient.
+  *   </li>
+  * </ul>
+  *
+  * @param knownCaseInsensitives Words known to appear with and without capitalized letters which help determine
+  * whether a span of text is contentful
+  * @param useLemmas If false, use the words of a sentence; if true, the lemmas
+  */
+// These knownCaseInsensitives are used to check for contentful spans.
+abstract class LexiconNER(val knownCaseInsensitives: Set[String], val useLemmas: Boolean) extends Tagger[String] with Serializable {
   /**
     * Matches the lexicons against this sentence
     * @param sentence The input sentence
@@ -21,38 +54,132 @@ abstract class LexiconNER(knownCaseInsensitives: Set[String], useLemmas: Boolean
     */
   def find(sentence: Sentence): Array[String]
   def getLabels: Seq[String]
-  def toString(stringBuilder: StringBuilder): Unit // This is to help with debugging.
 
-  protected def contentfulSpan(sentence: Sentence, start: Int, length: Int):Boolean = {
-    val (characters, letters, digits, upperCaseLetters, spaces) =
-      LexiconNER.scanText(sentence.words, start, start + length)
-    // a valid span must have letters > 0 and at least one of the other properties
-    val contentful = letters > 0 && (
-      digits > 0 ||
-      upperCaseLetters > 0 ||
-      spaces > 0 ||
-      characters > LexiconNER.KNOWN_CASE_INSENSITIVE_LENGTH ||
-      knownCaseInsensitives.contains(sentence.words(start))
-    )
+  /**
+    * The class is serializable and this method is used during testing to determine whether a reconstitued
+    * object is equal to the original without interfering with the operation of equals and getting into
+    * hash codes.  Is is not necessary for this operation to be efficient or complete.
+    * @param other The object to compare to
+    * @return Whether this and other are equal, at least as far is serialization is concerned
+    */
+  def equalsForSerialization(other: AnyRef): Boolean = {
+    other.isInstanceOf[LexiconNER] && {
+      val that = other.asInstanceOf[LexiconNER]
+
+      this.useLemmas == that.useLemmas &&
+      this.knownCaseInsensitives == that.knownCaseInsensitives
+    }
+  }
+
+  def hasCondition(wordsView: mutable.IndexedSeqView[String, Array[String]], condition: Char => Boolean): Boolean =
+    wordsView.exists(_.exists(condition))
+
+  def hasLetter(wordsView: mutable.IndexedSeqView[String, Array[String]]): Boolean =
+    hasCondition(wordsView, Character.isLetter)
+
+  def hasDigit(wordsView: mutable.IndexedSeqView[String, Array[String]]): Boolean =
+    hasCondition(wordsView, Character.isDigit)
+
+  def hasUpperCaseLetters(wordsView: mutable.IndexedSeqView[String, Array[String]]): Boolean =
+    hasCondition(wordsView, Character.isUpperCase)
+
+  def hasSpace(wordsView: mutable.IndexedSeqView[String, Array[String]]): Boolean = wordsView.length > 1
+
+  def countCharacters(wordsView: mutable.IndexedSeqView[String, Array[String]]): Int =
+    // Go ahead and calculate them all even though we only need to know if they exceed a value.
+    wordsView.foldLeft(0) { (sum, word) => sum + word.length }
+
+  val contentQualifiers: Array[mutable.IndexedSeqView[String, Array[String]] => Boolean] = Array(
+    // Start with the quick and easy ones.
+    hasSpace,
+    { wordsView => countCharacters(wordsView) > LexiconNER.KNOWN_CASE_INSENSITIVE_LENGTH },
+    hasDigit,
+    hasUpperCaseLetters,
+    { wordsView => knownCaseInsensitives.contains(wordsView.head) }
+  )
+
+  protected def contentfulSpan(sentence: Sentence, start: Int, length: Int): Boolean = {
+    val wordsView = sentence.words.view(start, start + length)
+    // A valid view/span must have a letter and at least one of the other qualifiers.
+    val contentful = hasLetter(wordsView) && contentQualifiers.exists(_(wordsView))
 
     contentful
   }
 
-  protected def getTokens(sentence: Sentence): Array[String] =
-      if (useLemmas) sentence.lemmas.get
-      else sentence.words
+  protected val getTokens: Sentence => Array[String] =
+    // Decide this once and for all and don't revisit it each time getTokens is called.
+    if (useLemmas) getLemmas
+    else getWords
+
+  protected def getLemmas(sentence: Sentence): Array[String] = sentence.lemmas.get
+
+  protected def getWords(sentence: Sentence): Array[String] = sentence.words
 }
 
-
 object LexiconNER {
-  // These are configuration values that affect object creation.
-  val USE_FAST = true // Otherwise, slow will be used
-  val USE_COMPACT = true // This applies to fast only
+  /**
+    * When true indicates use of the [[org.clulab.sequences.FastLexiconNERBuilder FastLexiconNERBuilder]]
+    * and otherwise the [[org.clulab.sequences.SlowLexiconNERBuilder SlowLexiconNERBuilder]]
+    * to construct the LexiconNER
+    */
+  var USE_FAST = true
+  /**
+    * If the [[org.clulab.sequences.FastLexiconNERBuilder FastLexiconNERBuilder]] is beind used,
+    * indicates when true that a [[org.clulab.sequences.CompactLexiconNER CompactLexiconNER]]
+    * should be created and otherwise a [[org.clulab.sequences.CombinedLexiconNER CombinedLexiconNER]]
+    */
+  var USE_COMPACT = true
   val USE_DEBUG = false // This applies to the slow one only
   val OVERRIDE_ENTITY_VALIDATOR = false // true is only for testing!
 
   val OUTSIDE_LABEL: String = "O"
   val KNOWN_CASE_INSENSITIVE_LENGTH: Int = 3 // this was tuned for Reach; if changed please rerun Reach unit tests
+
+  def something(kbs: Seq[String], overrideKBs: Option[Seq[String]], caseInsensitiveMatchings: Seq[Boolean], entityValidator: EntityValidator): Boolean = true
+  def something(b: Float, c: Boolean): Boolean = false
+
+  /** Create a LexiconNER from a pair of sequences of knowledge bases (KBs), the kbs and overrideKBs,
+    * with control over the case sensitivity of individual KBs via caseInsensitiveMatchings
+    *
+    * The matchings run parallel to the KBs.  That is, caseInsensitiveMatchings(n) is used for kbs(n).
+    * It is possible that contents of an overrideKB refers to a KB that does not exist.  In that
+    * situation, caseInsensitiveMatching is used as a fallback value.  With that, this method should
+    * encompass all the functionality of the other apply methods, which now feed into it.
+    *
+    * @param kbs KBs containing known entity names
+    * @param overrideKBs KBs containing override labels for entity names from kbs (necessary for the bio domain)
+    * @param caseInsensitiveMatchings case insensitivities corresponding to the kbs, matched by index
+    * @param entityValidator Filter which decides if a matched entity is valid
+    * @param lexicalVariationEngine Generates alternative spellings of an entity name (necessary for the bio domain)
+    * @param useLemmasForMatching If true, we use Sentence.lemmas instead of Sentence.words during matching
+    * @param defaultCaseInsensitive If true, tokens are matched case insensitively
+    * @return The new LexiconNER
+    */
+  def apply(kbs: Seq[String], overrideKBs: Option[Seq[String]], caseInsensitiveMatchings: Seq[Boolean],
+      entityValidator: EntityValidator, lexicalVariationEngine: LexicalVariations, useLemmasForMatching: Boolean,
+      defaultCaseInsensitive: Boolean): LexiconNER = {
+    val newEntityValidator =
+      if (OVERRIDE_ENTITY_VALIDATOR) EntityValidator.TRUE_VALIDATOR
+      else  entityValidator
+    val builder =
+      if (USE_FAST) new FastLexiconNERBuilder(USE_COMPACT)
+      else new SlowLexiconNERBuilder()
+
+    builder.build(kbs, caseInsensitiveMatchings, overrideKBs, newEntityValidator, lexicalVariationEngine,
+        useLemmasForMatching, defaultCaseInsensitive)
+  }
+
+  /** Same apply with some default values filled in */
+  def apply(kbs: Seq[String],
+            caseInsensitiveMatchings: Seq[Boolean],
+            entityValidator: EntityValidator,
+            useLemmasForMatching: Boolean): LexiconNER = {
+    apply(
+      kbs, None, caseInsensitiveMatchings,
+      entityValidator, new NoLexicalVariations, useLemmasForMatching,
+      false
+    )
+  }
 
   /**
     * Creates a LexiconNER from a list of KBs
@@ -68,21 +195,18 @@ object LexiconNER {
     * @param caseInsensitiveMatching If true, tokens are matched case insensitively
     * @return The new LexiconNER
     */
+
+  // Expand the caseInsensitiveMatching to plural and fill in parallel structure
+  // Do not initialize builder with the overall value.
+  // Pass the thing in the build arguments.
   def apply(kbs: Seq[String], overrideKBs: Option[Seq[String]],
       entityValidator: EntityValidator,
       lexicalVariationEngine: LexicalVariations,
       useLemmasForMatching: Boolean,
       caseInsensitiveMatching: Boolean): LexiconNER = {
-    val newEntityValidator =
-        if (OVERRIDE_ENTITY_VALIDATOR) EntityValidator.TRUE_VALIDATOR
-        else  entityValidator
-
-    if (USE_FAST)
-      new FastLexiconNERBuilder(caseInsensitiveMatching).build(kbs, overrideKBs, newEntityValidator,
-          lexicalVariationEngine, useLemmasForMatching)
-    else
-      new SlowLexiconNERBuilder(caseInsensitiveMatching).build(kbs, overrideKBs, newEntityValidator,
-          lexicalVariationEngine, useLemmasForMatching)
+    val caseInsensitiveMatchings = Array.fill(kbs.length)(caseInsensitiveMatching)
+    this(kbs, overrideKBs, caseInsensitiveMatchings, entityValidator, lexicalVariationEngine,
+        useLemmasForMatching, caseInsensitiveMatching)
   }
 
   /**
@@ -98,9 +222,9 @@ object LexiconNER {
     * @return The new LexiconNER
     */
   def apply(kbs: Seq[String],
-    entityValidator: EntityValidator = new TrueEntityValidator,
-    useLemmasForMatching: Boolean = false,
-    caseInsensitiveMatching: Boolean = false): LexiconNER = {
+      entityValidator: EntityValidator = new TrueEntityValidator,
+      useLemmasForMatching: Boolean = false,
+      caseInsensitiveMatching: Boolean = false): LexiconNER = {
     apply(kbs, None, entityValidator, new NoLexicalVariations, useLemmasForMatching, caseInsensitiveMatching)
   }
 
@@ -109,13 +233,11 @@ object LexiconNER {
     var letters = 0
     var digits = 0
     var upperCaseLetters = 0
-    val spaces = words.length - 1
+    val spaces = math.max(0, end - start - 1) // Spaces are between words, not after them.
 
-    for (offset <- start until end) {
-      val word = words(offset)
-      for (i <- word.indices) {
-        val c = word.charAt(i)
-        characters += 1
+    words.view(start, end).foreach { word =>
+      characters += word.length
+      word.foreach { c =>
         if (Character.isLetter(c)) letters += 1
         if (Character.isDigit(c)) digits += 1
         if (Character.isUpperCase(c)) upperCaseLetters += 1
@@ -124,273 +246,34 @@ object LexiconNER {
     (characters, letters, digits, upperCaseLetters, spaces)
   }
 
-  /** Merges labels from src into dst, without overlapping any existing labels in dst */
+  def isOutside(label: String): Boolean = label == OUTSIDE_LABEL
+  def isNotOutside(label: String): Boolean = !isOutside(label)
+
+  def countWhile(labels: Array[String], offset: Int, condition: String => Boolean): Int = {
+    var count = 0
+
+    while (offset + count < labels.length && condition(labels(offset + count)))
+      count += 1
+    count
+  }
+
+  /** Merges labels from src into dst without overlapping any existing labels in dst. */
   def mergeLabels(dst: Array[String], src: Array[String]) {
     assert(dst.length == src.length)
+    // Pick one or take the average of both.
+    val length = dst.length
+    // Skip over all the outside labels so that we will start notOutside.
+    var offset = countWhile(src, 0, isOutside)
 
-    var offset = 0
-    while (offset < dst.length) {
-      if (src(offset) != LexiconNER.OUTSIDE_LABEL) {
-        // no overlap allowed
-        // if overlap is detected, the corresponding labels in src are discarded
-        if (!overlap(dst, src, offset)) {
-          dst(offset) = src(offset)
-          offset += 1
-          while (offset < src.length && src(offset).startsWith("I-")) {
-            dst(offset) = src(offset)
-            offset += 1
-          }
-        }
-        else {
-          // overlap found; just ignore the labels in src
-          offset += 1
-          while (offset < src.length && src(offset).startsWith("I-")) {
-            offset += 1
-          }
-        }
-      }
-      else {
-        // nothing to merge
-        offset += 1
-      }
-    }
-  }
+    while (offset < length) {
+      val notOutsideCount = countWhile(src, offset, isNotOutside)
+      // Check that there is not anything in dst that should not be overwritten.
+      if (!dst.view(offset, offset + notOutsideCount).exists(isNotOutside))
+        src.copyToArray(dst, offset, notOutsideCount)
+      offset += notOutsideCount
 
-  // Used by mergeLabels above
-  private def overlap(dst: Array[String], src: Array[String], offset: Int): Boolean = {
-    var position = offset
-
-    if (dst(position) != LexiconNER.OUTSIDE_LABEL)
-      true
-    else {
-      position += 1
-      while (position < src.length && src(position).startsWith("I-")) {
-        if (dst(position) != LexiconNER.OUTSIDE_LABEL)
-          return true
-        position += 1
-      }
-      false
-    }
-  }
-}
-
-abstract class LexiconNERBuilder(caseInsensitiveMatching: Boolean) {
-  val logger: Logger = LoggerFactory.getLogger(classOf[LexiconNER])
-  val INTERN_STRINGS: Boolean = false
-
-  def build(kbs: Seq[String], overrideKBs: Option[Seq[String]],
-    entityValidator: EntityValidator, lexicalVariationEngine: LexicalVariations, useLemmasForMatching: Boolean): LexiconNER
-
-  protected def toJavaConsumer[T](consumer: T => Unit): Consumer[T] = {
-    new Consumer[T] {
-      override def accept(t: T): Unit = consumer(t)
-    }
-  }
-
-  protected def extractKBName(kb: String): String = {
-    val slash = kb.lastIndexOf("/")
-    val dot = kb.indexOf('.')
-    val name = kb.substring(slash + 1, dot)
-    name
-  }
-}
-
-class SlowLexiconNERBuilder(caseInsensitiveMatching: Boolean) extends LexiconNERBuilder(caseInsensitiveMatching) {
-
-  class BuildState(lexicalVariationEngine: LexicalVariations) {
-    val knownCaseInsensitives = new mutable.HashSet[String]()
-
-    def addWithLexicalVariations(matcher: BooleanHashTrie, tokens: Array[String]): Unit = {
-      // keep track of all lower case ents that are single token (necessary for case-insensitive matching)
-      if (tokens.length == 1) {
-        val token = tokens(0)
-        if (token.toLowerCase == token && caseInsensitiveMatching)
-          knownCaseInsensitives.add(tokens(0))
-      }
-      // add the original form
-      matcher.add(tokens)
-      // add the lexical variations
-      for (ts <- lexicalVariationEngine.lexicalVariations(tokens)) {
-        matcher.add(ts)
-      }
-    }
-  }
-
-  protected def newMatcher(label: String): BooleanHashTrie =
-    if (LexiconNER.USE_DEBUG)
-      new DebugBooleanHashTrie(label, caseInsensitive = caseInsensitiveMatching, internStrings = INTERN_STRINGS)
-    else
-      new BooleanHashTrie(label, caseInsensitive = caseInsensitiveMatching, internStrings = INTERN_STRINGS)
-
-  override def build(kbs: Seq[String], overrideKBs: Option[Seq[String]],
-      entityValidator: EntityValidator, lexicalVariationEngine: LexicalVariations, useLemmasForMatching: Boolean): SeparatedLexiconNER = {
-    logger.info("Beginning to load the KBs for the rule-based NER...")
-    val buildState = new BuildState(lexicalVariationEngine)
-    val overrideMatchers = getOverrideMatchers(overrideKBs, buildState) // first so they take priority during matching
-    val standardMatchers = getStandardMatchers(kbs, buildState)
-    logger.info("KB loading completed.")
-    new SeparatedLexiconNER((overrideMatchers ++ standardMatchers).toArray, buildState.knownCaseInsensitives.toSet,
-        useLemmasForMatching, entityValidator)
-  }
-
-  private def getStandardMatchers(kbs: Seq[String], buildState: BuildState): Seq[BooleanHashTrie] = {
-    def addLine(matcher: BooleanHashTrie, inputLine: String): Unit = {
-      val line = inputLine.trim
-      if (!line.startsWith("#")) {
-        val tokens = line.split("\\s+")
-        buildState.addWithLexicalVariations(matcher, tokens)
-      }
-    }
-
-    kbs.map { kb =>
-      val label = extractKBName(kb)
-      val matcher = newMatcher(label)
-
-      Serializer.using(loadStreamFromClasspath(kb)) { reader =>
-        reader.lines.forEach(toJavaConsumer[String] { line: String =>
-          addLine(matcher, line)
-        })
-      }
-      logger.info(s"Loaded matcher for label $label. The size of the first layer is ${matcher.entriesSize}.")
-      matcher
-    }
-  }
-
-  private def getOverrideMatchers(overrideKBs: Option[Seq[String]], buildState: BuildState): Seq[BooleanHashTrie] = {
-    // in these KBs, the name of the entity must be the first token, and the label must be the last
-    //   (tokenization is performed around TABs here)
-    def addLine(matchers: mutable.HashMap[String, BooleanHashTrie], inputLine: String): Unit = {
-      val line = inputLine.trim
-      if (!line.startsWith("#")) { // skip comments starting with #
-        val blocks = line.split("\t")
-        if (blocks.size >= 2) {
-          val entity = blocks.head // grab the text of the named entity
-          val label = blocks.last // grab the label of the named entity
-          val tokens = entity.split("\\s+")
-          val matcher = matchers.getOrElseUpdate(label, new BooleanHashTrie(label))
-
-          buildState.addWithLexicalVariations(matcher, tokens)
-        }
-      }
-    }
-
-    if (overrideKBs.isDefined) {
-      val matchers = new mutable.HashMap[String, BooleanHashTrie]()
-
-      overrideKBs.get.foreach { okb =>
-        Serializer.using(loadStreamFromClasspath(okb)) { reader =>
-          reader.lines.forEach(toJavaConsumer[String] { line =>
-            addLine(matchers, line)
-          })
-        }
-      }
-      for (name <- matchers.keySet.toList.sorted) {
-        val matcher = matchers(name)
-        logger.info(s"Loaded OVERRIDE matcher for label $name.  The size of the first layer is ${matcher.entriesSize}.")
-      }
-      matchers.values.toSeq
-    }
-    else Seq.empty[BooleanHashTrie]
-  }
-}
-
-class FastLexiconNERBuilder(caseInsensitiveMatching: Boolean) extends LexiconNERBuilder(caseInsensitiveMatching) {
-
-  class BuildState(lexicalVariationEngine: LexicalVariations) {
-    val intHashTrie = new IntHashTrie(caseInsensitiveMatching)
-    val labelToIndex = new mutable.HashMap[String, Int] // This just maps the labels to an integer
-    val knownCaseInsensitives = new mutable.HashSet[String]()
-
-    def getCount: Int = intHashTrie.entriesSize
-
-    protected def add(label: String, tokens: Array[String]): Unit = {
-      val index = labelToIndex.getOrElseUpdate(label, labelToIndex.size)
-
-      intHashTrie.add(tokens, index)
-    }
-
-    def addWithLexicalVariations(label: String, tokens: Array[String]): Unit = {
-      // keep track of all lower case ents that are single token (necessary for case-insensitive matching)
-      if (tokens.length == 1) {
-        val token = tokens(0)
-        if (token.toLowerCase == token && caseInsensitiveMatching)
-          knownCaseInsensitives.add(tokens(0))
-      }
-      // add the original form
-      add(label, tokens)
-      // add the lexical variations
-      for (ts <- lexicalVariationEngine.lexicalVariations(tokens)) {
-        add(label, ts)
-      }
-    }
-  }
-
-  protected def newMatcher(label: String) =
-      new IntHashTrie(caseInsensitive = caseInsensitiveMatching, internStrings = INTERN_STRINGS)
-
-  override def build(kbs: Seq[String], overrideKBs: Option[Seq[String]],
-      entityValidator: EntityValidator, lexicalVariationEngine: LexicalVariations, useLemmasForMatching: Boolean): LexiconNER = {
-    logger.info("Beginning to load the KBs for the rule-based NER...")
-    val buildState = new BuildState(lexicalVariationEngine)
-    addOverrideKBs(overrideKBs, buildState) // first so they take priority during matching
-    addStandardKBs(kbs, buildState)
-    logger.info("KB loading completed.")
-
-    val labels = buildState.labelToIndex.toArray.sortBy(_._2).map(_._1)
-
-    if (LexiconNER.USE_COMPACT)
-      CompactLexiconNER(buildState.intHashTrie, labels, buildState.knownCaseInsensitives.toSet, useLemmasForMatching, entityValidator)
-    else
-      new CombinedLexiconNER(buildState.intHashTrie, labels, buildState.knownCaseInsensitives.toSet, useLemmasForMatching, entityValidator)
-  }
-
-  private def addStandardKBs(kbs: Seq[String], buildState: BuildState): Unit = {
-    def addLine(inputLine: String, label: String): Unit = {
-      val line = inputLine.trim
-      if (!line.startsWith("#")) {
-        val tokens = line.split("\\s+")
-        buildState.addWithLexicalVariations(label, tokens)
-      }
-    }
-
-    kbs.foreach { kb =>
-      val label = extractKBName(kb)
-      val beforeCount = buildState.getCount
-      Serializer.using(loadStreamFromClasspath(kb)) { reader =>
-        reader.lines.forEach(toJavaConsumer[String] { line: String =>
-          addLine(line, label)
-        })
-      }
-      var afterCount = buildState.getCount
-      logger.info(s"Loaded matcher for label $label. The number of entries added to the first layer was ${afterCount - beforeCount}.")
-    }
-  }
-
-  private def addOverrideKBs(overrideKBs: Option[Seq[String]], buildState: BuildState): Unit = {
-    // in these KBs, the name of the entity must be the first token, and the label must be the last
-    //   (tokenization is performed around TABs here)
-    def addLine(inputLine: String): Unit = {
-      val line = inputLine.trim
-      if (!line.startsWith("#")) { // skip comments starting with #
-        val blocks = line.split("\t")
-        if (blocks.size >= 2) {
-          val entity = blocks.head   // grab the text of the named entity
-          val label = blocks.last    // grab the label of the named entity
-          val tokens = entity.split("\\s+")
-          buildState.addWithLexicalVariations(label, tokens)
-        }
-      }
-    }
-
-    overrideKBs.getOrElse(Seq.empty).foreach { okb =>
-      val beforeCount = buildState.getCount
-      Serializer.using(loadStreamFromClasspath(okb)) { reader =>
-        reader.lines.forEach(toJavaConsumer[String] { line: String =>
-          addLine(line)
-        })
-      }
-      var afterCount = buildState.getCount
-      logger.info(s"Loaded OVERRIDE matchers for all labels.  The number of entries added to the first layer was ${afterCount - beforeCount}.")
+      val outsideCount = countWhile(src, offset, isOutside)
+      offset += outsideCount
     }
   }
 }