+++ CHANGES.md
@@ -7,6 +7,8 @@
 + **8.3.0** - Supply InputStreamer to aid in reading materials from resources and files, possibly after serialization.
 + **8.3.0** - Modernize build with separate sbt files, improved scoping, updated plugins, and most recent versions of Scala 11 and 12.
 + **8.3.0** - Update Jenkins file for memory, Scala version, cross testing.
++ **8.3.0** - CluCoreProcessor now supports a custom LexiconNER constructed on the fly 
++ **8.3.0** - Allow LexiconNER collection to use KBs with differing case sensitivity
 + **8.2.6** - Bug fix: handle cycles in dependency trees produced by CoreNLPProcessor.
 + **8.2.5** - Added CluCoreProcessor, which uses all our tools + Stanford's numeric entity recognizer. 
 + **8.2.5** - Update to FatDynet 0.3.0 in serial mode. This fixes memory leaks and crashed in the DyNet Scala wrapper.